improved code structure
