+++ .gitignore
@@ -6,3 +6,4 @@ logs
 .idea
 .DS_Store
 build
+data
