--- src/template.csv
@@ -1,3 +1,0 @@
-no,first_name,last_name,username,active,email,role,place,password
-1,Frank,Scarlet,stiffler,1,stiffler@gmail.com,cha,"601751, 900200",P@ssword
-2,Ichigo,Kurasaki,ichigo,1,ichigo@gmail.com,cha,"601751",P@ssword
