+++ src/types/auth/index.ts
@@ -1,17 +1,17 @@
 /**
- * Models for auth on Superset
+ * Authentication related types
  */
 
 export interface LoginRequest {
   username: string;
   password: string;
   provider: string;
-  refresh: boolean
+  refresh: boolean;
 }
 
 export interface LoginResponse {
   access_token: string;
-  refresh_token: string
+  refresh_token: string;
 }
 
 export interface CSRFResponse {
